+++ lib/EasyPost/Constant/Constants.php
@@ -43,7 +43,6 @@ abstract class Constants
 
     const NO_USER_FOUND_ERROR = 'No user found with the given ID.';
     const NO_RESPONSE_ERROR = 'Did not receive a response from %s.';
-    const SEND_STRIPE_DETAILS_ERROR = 'Could not send card details to Stripe, please try again later.';
     const UNDEFINED_PROPERTY_ERROR = 'EasyPost Notice: Undefined property of %s instance: %s';
     const NO_MATCHING_MOCK_REQUEST = 'No matching mock request found for %s %s';
     const END_OF_PAGINATION = 'There are no more pages to retrieve.';